+++ lib/mix/tasks/scapa.gen.versions.ex
@@ -23,7 +23,7 @@ defmodule Mix.Tasks.Scapa.Gen.Versions do
     {parsed, _argv, _errors} =
       OptionParser.parse(argv, aliases: [c: :config_file], strict: [config_file: :string])
 
-    config = Config.fetch_config(parsed[:config_file])
+    config = Config.fetch_config(parsed[:config_file] || Config.default_config_path())
 
     case Scapa.CLI.generate_versions(config) do
       {:ok, sync} -> override_files(sync)